FastAPI/app/requirement.txt
certifi==2024.8.30
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.15.4
idna==3.8
numpy==2.1.0
pandas==2.2.2
platformdirs==4.2.2
pyFirmata==1.1.0
pyjokes==0.8.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
pydantic
typing
mysql-connector-python
sqlalchemy
mysqlclient
fastapi
emailvalidator
passlib
Bcrypt
python-jose
cryptography